refacotr: rename @kubricate/kubernetes to @kubricate/plugin-kubernetes across the codebase
