+++ .changeset/config.json
@@ -2,7 +2,7 @@
   "$schema": "https://unpkg.com/@changesets/config@3.1.1/schema.json",
   "changelog": "@changesets/cli/changelog",
   "commit": true,
-  "fixed": [["kubricate", "@kubricate/core", "@kubricate/plugin-env", "@kubricate/stacks", "@kubricate/kubernetes"]],
+  "fixed": [["kubricate", "@kubricate/core", "@kubricate/plugin-env", "@kubricate/stacks", "@kubricate/plugin-kubernetes"]],
   "linked": [],
   "access": "public",
   "baseBranch": "main",
