+++ examples/with-secret/package.json
@@ -32,7 +32,7 @@
   "dependencies": {
     "@kubricate/core": "workspace:*",
     "@kubricate/plugin-env": "workspace:*",
-    "@kubricate/kubernetes": "workspace:*",
+    "@kubricate/plugin-kubernetes": "workspace:*",
     "@kubricate/stacks": "workspace:*",
     "kubernetes-models": "^4.4.2"
   }
