+++ examples/with-secret/src/setup-secrets.ts
@@ -1,5 +1,5 @@
 import { SecretManager } from '@kubricate/core';
-import { OpaqueSecretProvider, DockerConfigSecretProvider } from '@kubricate/kubernetes';
+import { OpaqueSecretProvider, DockerConfigSecretProvider } from '@kubricate/plugin-kubernetes';
 import { EnvConnector } from '@kubricate/plugin-env';
 
 export const secretManager = new SecretManager()
